[cli] make linkmetrics cli as a separate module (#9619)
