+++ src/cli/BUILD.gn
@@ -49,6 +49,8 @@ openthread_cli_sources = [
   "cli_history.hpp",
   "cli_joiner.cpp",
   "cli_joiner.hpp",
+  "cli_link_metrics.cpp",
+  "cli_link_metrics.hpp",
   "cli_mac_filter.cpp",
   "cli_mac_filter.hpp",
   "cli_network_data.cpp",
