+++ src/cli/CMakeLists.txt
@@ -42,6 +42,7 @@ set(COMMON_SOURCES
     cli_dns.cpp
     cli_history.cpp
     cli_joiner.cpp
+    cli_link_metrics.cpp
     cli_mac_filter.cpp
     cli_network_data.cpp
     cli_output.cpp
